fix each method
